requirements.txt
asgiref==3.8.1
asttokens==3.0.0
attrs==25.4.0
certifi==2025.6.15
decorator==5.2.1
distlib==0.3.9
Django==5.2.2
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
executing==2.2.0
filelock==3.18.0
inflection==0.5.1
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
matplotlib-inline==0.1.7
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==12.1.0
pipenv==2025.0.3
platformdirs==4.3.8
prompt_toolkit==3.0.51
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
PyJWT==2.11.0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
setuptools==80.9.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
uritemplate==4.2.0
virtualenv==20.31.2
wcwidth==0.2.13
